Added a `contract` argument the TransferChain to allow transfer between airdrop contracts Fixed a bug where all sequence responses were Int rather than Long values

+++ eos-http-rpc/src/main/kotlin/com/memtrip/eos/http/rpc/model/history/request/GetActions.kt
@@ -20,6 +20,6 @@ import com.squareup.moshi.JsonClass
 @JsonClass(generateAdapter = true)
 data class GetActions(
     val account_name: String,
-    val pos: Int? = null,
-    val offset: Int? = null
+    val pos: Long? = null,
+    val offset: Long? = null
 )
